backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
python-multipart==0.0.6
torch==2.1.0
transformers==4.35.0
diffusers==0.24.0
opencv-python==4.8.1.78
moviepy==1.0.3
pillow==10.1.0
numpy==1.24.3
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.25.1
xformers==0.0.22
PyJWT==2.8.0
bcrypt==4.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
ffmpeg-python==0.2.0
pydub==0.25.1
locust==2.17.0
pytest-benchmark==4.0.0
psutil==5.9.6
